requirements.txt
streamlit
pandas
altair
plotly[express]
matplotlib
network
numpy
scikit-learn
nltk
textblob